refactor(logic): split gp_old_world_resource_redistribution into part-of concern files (#3290)
Phase 0 file decomposition of the 564-non-comment-line setup module into
two `part of` fragments by concern (tile-scan/fairness helpers and
quota/placement logic), keeping the public entrypoints and infeasible
exception in the library file. Behaviour-preserving: shared library scope
(imports, private helpers, visibility unchanged).

Refs #3290
